BACKPORT: staging: android: binder: move to the real part of the kernel
The Android binder code has been "stable" for many years now.  No matter
what comes in the future, we are going to have to support this API, so
might as well move it to the "real" part of the kernel as there's no
real work that needs to be done to the existing code.

[@dev-harsh1998]: Adjust for 3.10 kernel

Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: dev-harsh1998 <harshitjain6751@gmail.com>
Change-Id: I5b8dc7954a0f5117f414fc4db8a29502f8885788
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>