[BACKPORT]: ANDROID: binder: Add BINDER_GET_NODE_INFO_FOR_REF ioctl.
This allows the context manager to retrieve information about nodes
that it holds a reference to, such as the current number of
references to those nodes.

Such information can for example be used to determine whether the
servicemanager is the only process holding a reference to a node.
This information can then be passed on to the process holding the
node, which can in turn decide whether it wants to shut down to
reduce resource usage.

[@dev-harsh1998]: Adjust for our backport.

Signed-off-by: Martijn Coenen <maco@android.com>
Signed-off-by: dev-harsh1998 <dev-harsh1998@hotmail.com>
Change-Id: Ic1b56c3a666ebf8f16e1c0dc8addb44728b4f606
Signed-off-by: Mrinal Ghosh <mg712702@gmail.com>